backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.2.1
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.22
email-validator==2.3.0
openai==1.99.9
reportlab==4.4.10
pillow==12.1.1
aiofiles==25.1.0
